+++ sdk/lib/io/io_sink.dart
@@ -15,8 +15,7 @@ part of dart.io;
  * While a stream is being added using [addStream], any further attempts
  * to add or write to the [IOSink] will fail until the [addStream] completes.
  *
- * If data is added to the [IOSink] after the sink is closed, the data will be
- * ignored. Use the [done] future to be notified when the [IOSink] is closed.
+ * It is an error to add data to the [IOSink] after the sink is closed.
  */
 abstract class IOSink implements StreamSink<List<int>>, StringSink {
   /**
@@ -148,11 +147,12 @@ class _StreamSinkImpl<T> implements StreamSink<T> {
   _StreamSinkImpl(this._target);
 
   void add(T data) {
-    if (_isClosed) return;
+    if (_isClosed) throw new StateError("StreamSink is closed");
     _controller.add(data);
   }
 
   void addError(error, [StackTrace stackTrace]) {
+    if (_isClosed) throw new StateError("StreamSink is closed");
     _controller.addError(error, stackTrace);
   }