+++ tests/standalone/io/socket_exception_test.dart
@@ -90,6 +90,8 @@ class SocketExceptionTest {
         }
         Expect.isFalse(exceptionCaught);
         Expect.isFalse(wrongExceptionCaught);
+
+        // From here exceptions are expected.
         try {
           List<int> buffer = new List<int>(10);
           client.add(buffer);
@@ -98,10 +100,8 @@ class SocketExceptionTest {
         } catch (ex) {
           wrongExceptionCaught = true;
         }
-        Expect.isFalse(exceptionCaught);
+        Expect.isTrue(exceptionCaught);
         Expect.isFalse(wrongExceptionCaught);
-
-        // From here exceptions are expected.
         exceptionCaught = false;
         try {
           client.port;